refactor: remove initial implementation
