+++ src/instruction_builder/mod.rs
@@ -11,7 +11,6 @@ mod init_validator_fees_vault;
 mod protocol_claim_fees;
 mod top_up_ephemeral_balance;
 mod undelegate;
-mod undelegate_ephemeral_balance;
 mod validator_claim_fees;
 mod whitelist_validator_for_program;
 
@@ -27,6 +26,5 @@ pub use init_validator_fees_vault::*;
 pub use protocol_claim_fees::*;
 pub use top_up_ephemeral_balance::*;
 pub use undelegate::*;
-pub use undelegate_ephemeral_balance::*;
 pub use validator_claim_fees::*;
 pub use whitelist_validator_for_program::*;
