+++ src/processor/close_ephemeral_balance.rs
@@ -1,8 +1,12 @@
 use crate::ephemeral_balance_seeds_from_payer;
-use crate::processor::utils::loaders::{load_initialized_pda, load_signer};
-use crate::processor::utils::pda::close_pda;
+use crate::processor::utils::loaders::{load_pda, load_signer};
+use solana_program::msg;
+use solana_program::program::invoke_signed;
 use solana_program::program_error::ProgramError;
-use solana_program::{account_info::AccountInfo, entrypoint::ProgramResult, pubkey::Pubkey};
+use solana_program::system_instruction::transfer;
+use solana_program::{
+    account_info::AccountInfo, entrypoint::ProgramResult, pubkey::Pubkey, system_program,
+};
 
 /// Process the closing of an ephemeral balance account
 ///
@@ -10,6 +14,7 @@ use solana_program::{account_info::AccountInfo, entrypoint::ProgramResult, pubke
 ///
 /// 0: `[signer]` payer to pay for the transaction and receive the refund
 /// 1: `[writable]` ephemeral balance account we are closing
+/// 2: `[]` the system program
 ///
 /// Requirements:
 ///
@@ -27,21 +32,45 @@ pub fn process_close_ephemeral_balance(
     let index = *data.first().ok_or(ProgramError::InvalidInstructionData)?;
 
     // Load Accounts
-    let [payer, ephemeral_balance_account] = accounts else {
+    let [payer, ephemeral_balance_account, system_program] = accounts else {
         return Err(ProgramError::NotEnoughAccountKeys);
     };
 
     load_signer(payer, "payer")?;
 
-    load_initialized_pda(
+    let ephemeral_balance_seeds: &[&[u8]] = ephemeral_balance_seeds_from_payer!(payer.key, index);
+    let ephemeral_balance_bump = load_pda(
         ephemeral_balance_account,
-        ephemeral_balance_seeds_from_payer!(payer.key, index),
+        ephemeral_balance_seeds,
         &crate::id(),
         true,
         "ephemeral balance",
     )?;
+    if ephemeral_balance_account.owner != &system_program::id() {
+        msg!(
+            "ephemeral balance expected to be owned by system program: {}",
+            system_program::id()
+        );
+        return Err(ProgramError::InvalidAccountOwner);
+    }
 
-    close_pda(ephemeral_balance_account, payer)?;
+    let amount = ephemeral_balance_account.lamports();
+    if amount == 0 {
+        return Ok(());
+    }
+
+    let ephemeral_balance_bump_slice: &[u8] = &[ephemeral_balance_bump];
+    let ephemeral_balance_signer_seeds =
+        [ephemeral_balance_seeds, &[ephemeral_balance_bump_slice]].concat();
+    invoke_signed(
+        &transfer(ephemeral_balance_account.key, payer.key, amount),
+        &[
+            ephemeral_balance_account.clone(),
+            payer.clone(),
+            system_program.clone(),
+        ],
+        &[&ephemeral_balance_signer_seeds],
+    )?;
 
     Ok(())
 }
